Additon of accounting navbar entry
