+++ frontend/src/app/navbar/navbar.component.spec.ts
@@ -29,6 +29,7 @@ import { MatDialogModule } from '@angular/material/dialog'
 import { MatDividerModule } from '@angular/material/divider'
 import { MatGridListModule } from '@angular/material/grid-list'
 import { NgMatSearchBarModule } from 'ng-mat-search-bar'
+import { AdminGuard } from '../app.routing'
 
 class MockSocket {
   on (str: string, callback) {
@@ -48,6 +49,7 @@ describe('NavbarComponent', () => {
   let mockSocket
   let socketIoService
   let location
+  let adminGuard
 
   beforeEach(async(() => {
 
@@ -67,6 +69,8 @@ describe('NavbarComponent', () => {
     mockSocket = new MockSocket()
     socketIoService = jasmine.createSpyObj('SocketIoService', ['socket'])
     socketIoService.socket.and.returnValue(mockSocket)
+    adminGuard = jasmine.createSpyObj('AdminGuard',['tokenDecode'])
+    adminGuard.tokenDecode.and.returnValue(of(true))
 
     TestBed.configureTestingModule({
       declarations: [ NavbarComponent, SearchResultComponent ],
@@ -101,6 +105,7 @@ describe('NavbarComponent', () => {
         { provide: ChallengeService, useValue: challengeService },
         { provide: CookieService, useValue: cookieService },
         { provide: SocketIoService, useValue: socketIoService },
+        { provide: AdminGuard, useValue: adminGuard },
         TranslateService
       ]
     })
