+++ frontend/src/app/roles.ts
@@ -1,0 +1,6 @@
+export const roles = {
+  customer: 'customer',
+  prime: 'prime',
+  accounting: 'accounting',
+  admin: 'admin'
+}
